Extract BlurEffect interface
