+++ haze/src/skikoMain/kotlin/dev/chrisbanes/haze/HazeSourceNode.skiko.kt
@@ -3,4 +3,11 @@
 
 package dev.chrisbanes.haze
 
+import androidx.compose.ui.graphics.drawscope.DrawScope
+
 internal actual fun HazeSourceNode.clearHazeAreaLayerOnStop() = Unit
+
+internal actual fun HazeEffectNode.selectBlurEffect(drawScope: DrawScope): BlurEffect = when {
+  blurEnabled -> RenderEffectBlurEffect
+  else -> ScrimBlurEffect
+}
